delete rezip()
